requirements.txt
# Google APIs
google-genai>=1.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
apify-client>=2.4.1

# Email Verification
dnspython>=2.4.0
email-validator>=2.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=9.1.4

# Notifications (optional)
python-telegram-bot>=20.0
